requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.8
python-dotenv>=1.0
qdrant-client>=1.9
langchain>=0.3
langgraph>=0.2
langchain-community>=0.3
langchain-openai>=0.2
tiktoken>=0.7
numpy>=1.26
pypdf>=4.2
langchain-text-splitters>=0.3
langchain-ollama>=0.1
